Use HashMap in MapsFragment.java
To reduce redundant code, a HashMap is used with a Marker as its key and a String for its value. The value in the HashMap is the name of the xml file that belongs to the location. We can use the name of the xml file to pass into the popupViewCreator, achieving the same functionality without the use of a switch statement.